refactor: architecture review - unify, simplify, delete code
Round 1-10 findings:

Module Structure:
- Moved paint.rs, paint_builders.rs, paint_tests.rs into paint/ directory
- Moved replay.rs and replay_ops/ into replay/ directory
- Moved events.rs, events_message.rs, events_turn.rs into events/ subdirectory
- Removed dead tui/update/error.rs module

Code Simplification:
- Removed duplicate OnboardingStep enum (now single source in components/onboarding)
- Removed unnecessary convert_onboarding_step function
- Simplified model provider functions (removed redundant sort_by calls)
- Removed unused eq_data_variant duplicate implementation in Msg

Linter Compliance:
- Fixed render_plan_frame (41 lines) by splitting into smaller helpers
- Made find_most_recent_spinner_index and get_msg_type private
- Fixed deprecated ratatui API (get_mut -> cell_mut)
- Added Debug derives to various structs

Bug Fixes:
- Fixed TextAreaBookmark -> CursorMove::End
- Fixed rerank -> update_filtered method visibility
- Fixed GitInfo field access
- Simplified OverlayViewModel builder
- Fixed braille_frame type casting
- Fixed TuiMode::Home -> TuiMode::HomeScreen

+++ crates/runie-tui/src/components/agent_list/builder.rs
@@ -9,7 +9,7 @@ pub(crate) struct AgentListBuilder {
     tokens: u64,
     cost: f64,
     agent_running: bool,
-    braille_frame: usize,
+    braille_frame: u8,
 }
 
 impl AgentListBuilder {
@@ -61,7 +61,7 @@ impl AgentListBuilder {
     }
 
     pub(crate) fn braille_frame(mut self, frame: usize) -> Self {
-        self.braille_frame = frame;
+        self.braille_frame = frame as u8;
         self
     }
